backend-v2/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
psycopg2-binary==2.9.9
pgvector==0.2.5
sqlalchemy==2.0.30
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
pydicom==2.4.4
requests==2.32.3
python-multipart==0.0.9
python-dotenv==1.0.1
reportlab==4.2.0
# RAG Engine dependencies
sentence-transformers>=2.7.0
rank-bm25>=0.2.2
numpy>=1.26.0
# NL2SQL Gemini fallback (optional)
google-generativeai>=0.7.0
